Move helpers into `CommunityHelpers`
Move helper methods specific to `EnvelopeCommunities` into
`CommunityHelpers`.
